+++ src/main/java/bowling/domain/FrameLast.java
@@ -1,4 +1,4 @@
-package refactor;
+package bowling.domain;
 
 import java.util.ArrayList;
 import java.util.List;
