+++ src/main/java/bowling/domain/Frames.java
@@ -1,4 +1,4 @@
-package refactor;
+package bowling.domain;
 
 import java.util.List;
 import java.util.Objects;
