+++ src/main/java/bowling/view/Output.java
@@ -1,4 +1,9 @@
-package refactor;
+package bowling.view;
+
+import bowling.domain.Frames;
+import bowling.domain.Player;
+import bowling.domain.State;
+import bowling.domain.Subtotal;
 
 import java.util.List;
